Added line that shows how to remake chart on change
The line is commented out, since uncautious editing of the input
fields can cause the chart to be remade with nonsensical parameters,
and causing the simulation to be re-run every time there's an edit
in one of the fields is kind of crazy.  But it shows Reagent in action,
and shows how to re-make a d3 object (not how to modify it) in response
to modification in a Reagent form.